Refactoring code as per Shea's advice
