+++ client/src/components/App.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react';
-import StudentSessionView from './StudentSessionView';
-import './App.css';
+import StudentSessionView from './pages/StudentSessionView';
+import '../styles/App.css';
 
 class App extends React.Component {
   render() {
