+++ client/src/components/includes/SessionPopularQuestionsContainer.tsx
@@ -1,18 +1,18 @@
 import * as React from 'react';
-import './SessionPopularQuestionsContainer.css';
+import '../../styles/SessionPopularQuestionsContainer.css';
 import SessionPopularQuestion from './SessionPopularQuestion';
 
 class SessionPopularQuestionsContainer extends React.Component {
     render() {
-        return(
+        return (
             <div className="SessionPopularQuestionsContainer">
                 <div className="SessionPopularQuestionsContainer-Title">
                     Popular questions <a className="SessionPopularQuestionsContainer-Collapse">(collapse)</a>
                 </div>
                 <div className="SessionPopularQuestionsContainer-List">
                     <SessionPopularQuestion question="This is a sample question" numPeople={6} />
-                    <SessionPopularQuestion question="This is another question" numPeople={3}/>
-                    <SessionPopularQuestion question="And one more question" numPeople={2}/>
+                    <SessionPopularQuestion question="This is another question" numPeople={3} />
+                    <SessionPopularQuestion question="And one more question" numPeople={2} />
                 </div>
             </div>
         );
