+++ client/src/components/includes/registerServiceWorker.ts
