+++ client/src/components/pages/StudentSessionView.tsx
@@ -1,12 +1,12 @@
 import * as React from 'react';
-import './StudentSessionView.css';
-import SessionInformationHeader from './SessionInformationHeader';
-import SessionQuestionsContainer from './SessionQuestionsContainer';
-import SessionPopularQuestionsContainer from './SessionPopularQuestionsContainer';
+import '../../styles/StudentSessionView.css';
+import SessionInformationHeader from '../includes/SessionInformationHeader';
+import SessionQuestionsContainer from '../includes/SessionQuestionsContainer';
+import SessionPopularQuestionsContainer from '../includes/SessionPopularQuestionsContainer';
 
 class StudentSessionView extends React.Component {
     render() {
-        return(
+        return (
             <div className="StudentSessionView">
                 <SessionInformationHeader />
                 <SessionPopularQuestionsContainer />
