+++ client/src/styles/SessionInformationHeader.css
