+++ client/src/styles/StudentSessionView.css
